+++ packages/lib/src/constants/runtime.ts
@@ -5,3 +5,5 @@ export const FEDERATION_SHARED_PREFIX = '__federation_shared_';
 export const VITE_BASE_PLACEHOLDER = '__VITE_BASE_PLACEHOLDER__';
 export const VITE_ASSETS_DIR_PLACEHOLDER = '__VITE_ASSETS_DIR_PLACEHOLDER__';
 export const CJS_EXPORTS_RE = /exports\.([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=/g;
+
+export const FEDERATION_DEPS_DIR = '.federation-deps';
