+++ packages/lib/src/constants/virtual-modules.ts
@@ -4,3 +4,6 @@ export const VIRTUAL_FN_IMPORT = '__federation_fn_import';
 export const VIRTUAL_FN_IMPORT_RESOLVED = `\0virtual:${VIRTUAL_FN_IMPORT}`;
 export const VIRTUAL_FN_SATISFY = '__federation_fn_satisfy';
 export const REMOTE_ENTRY_HELPER_PREFIX = '__remoteEntryHelper__';
+
+export const SHARED_VIRTUAL_PREFIX = 'virtual:__federation_shared__:';
+export const RESOLVED_SHARED_PREFIX = '\0' + SHARED_VIRTUAL_PREFIX;
