+++ packages/lib/src/dev/cors.ts
@@ -1,0 +1,20 @@
+import type { IncomingMessage, ServerResponse } from 'node:http';
+
+/** Add CORS headers so the HOST browser can load files from this remote. */
+export const corsMiddleware = (
+  req: IncomingMessage,
+  res: ServerResponse,
+  next: () => void
+): void => {
+  res.setHeader('Access-Control-Allow-Origin', '*');
+  res.setHeader('Access-Control-Allow-Methods', '*');
+  res.setHeader('Access-Control-Allow-Headers', '*');
+
+  if (req.method === 'OPTIONS') {
+    res.statusCode = 204;
+    res.end();
+    return;
+  }
+
+  next();
+};
